backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
psycopg[binary]==3.1.18
greenlet==3.0.3
pandas==2.1.3
numpy>=1.25.0
openpyxl==3.1.2
celery==5.3.6
redis==5.0.1
structlog==24.1.0
langchain==0.1.16
langchain-core==0.1.45
langgraph==0.0.30
langchain-google-genai 
langchain-groq 
google-generativeai==0.4.1
openai==1.16.2
anthropic==0.25.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
